changed sitter.user_id to sitter.userID in sitterList and moved state around so OwnerDashboard is owns the sitterList state
