+++ src/components/SitterList.js
@@ -1,40 +1,11 @@
-import React, {useState, useEffect} from 'react';
-import {Alert, Table} from 'react-bootstrap';
+import React from 'react';
+import {Table} from 'react-bootstrap';
 import Moment from 'moment';
-import { useAuth } from '../contexts/AuthContext';
 import {Link} from 'react-router-dom';
-import axios from 'axios';
-import { Button } from 'react-bootstrap';
 
 import './SitterList.css'
-import SitterMap from './SitterMap';
 
-const SitterList = ({ baseURL, currentUserData }) => {
-    const { currentUser } = useAuth();
-    const [sitterList, setSitterList] = useState(null);
-    const [error, setError] = useState({variant: '', message: ''});
-    const [showMap, setShowMap] = useState(false);
-
-    useEffect(()=>{
-        axios.get(baseURL + '/sitters')
-            .then((response) => {
-                const apiSitterList = Object.values(response.data)
-                if (Object.keys(response.data)[0] !== 'message') {
-                    const userIDs = Object.keys(response.data)
-                    for(let i in userIDs) {
-                        apiSitterList[i].user_id = userIDs[i];
-                    }
-                    setSitterList(apiSitterList)
-                } else {
-                    setError({variant: 'warning', message: Object.values(response.data)[0]})
-                }
-            })
-            .catch((error) => {
-                const message=`There was an error with your request. ${error.message}.`;
-                setError({variant: 'danger', message: message});
-                console.log(message);
-            })
-    }, [baseURL])
+const SitterList = ({ sitterList, currentUserData }) => {
 
     const showSitterList = () => {
         return(
@@ -50,31 +21,31 @@ const SitterList = ({ baseURL, currentUserData }) => {
                 </thead>
                 <tbody>
                     {(sitterList).map((sitter) => {
-                        if (sitter.auth_id !== currentUser.uid) {
+                        if (sitter.auth_id !== currentUserData.uid) {
                             return(
-                                <tr key={sitter.user_id}>
+                                <tr key={sitter.userID}>
                                     <td>
-                                        <Link to={`/users/${sitter.user_id}`}>
+                                        <Link to={`/users/${sitter.userID}`}>
                                             {sitter.full_name}
                                         </Link>
                                     </td>
                                     {/* <td>
-                                        <Link to={`/users/${sitter.user_id}`}>
+                                        <Link to={`/users/${sitter.userID}`}>
                                             {sitter.username}
                                         </Link>
                                     </td> */}
                                     <td>
-                                        <Link to={`/users/${sitter.user_id}`}>
+                                        <Link to={`/users/${sitter.userID}`}>
                                             {sitter.phone_number}
                                         </Link>
                                     </td>
                                     <td>
-                                        <Link to={`/users/${sitter.user_id}`}>
+                                        <Link to={`/users/${sitter.userID}`}>
                                             {Moment(sitter.date_joined).format('MM-DD-YYYY')}
                                         </Link>
                                     </td>
                                     <td>
-                                        <Link to={`/users/${sitter.user_id}`}>
+                                        <Link to={`/users/${sitter.userID}`}>
                                             { sitter.rating ? sitter.rating : 'N/A'}
                                         </Link>
                                     </td>
@@ -89,24 +60,12 @@ const SitterList = ({ baseURL, currentUserData }) => {
         )
     }
 
-    const onViewMapClick = () => {
-        setShowMap(!showMap);
-    }
-
     if (!sitterList) {
         return <div></div>;
     }
 
     return (
         <div className='sitter-list'>
-            { error.message && <Alert variant={error.variant}>{error.message}</Alert>} 
-            { showMap ? 
-                <div>
-                    <Button variant='outline-secondary' onClick={onViewMapClick}>Hide Map</Button>
-                    <SitterMap sitterList={sitterList} currentUserData={currentUserData} />
-                </div> 
-            : <Button variant='outline-secondary' onClick={onViewMapClick}>View Map</Button>
-            }            
             {showSitterList()}
         </div>
     )
