+++ src/components/SitterMap.js
@@ -27,7 +27,7 @@ const SitterMap = ({ sitterList, currentUserData }) => {
     const loadSitterListMarkers = () => {
         const apiSitterCoords = []
         sitterList.forEach((sitter) => {
-            if (sitter.user_id !== currentUserData.userID) {
+            if (sitter.userID !== currentUserData.userID) {
                 axios.get(createGeocodeURL(sitter))
                     .then((response) =>{
                         const apiSitter = {
